feat: Move utils methods in CozyUtils
This is useful to use these methods froms apps
