+++ src/CozyUtils.js
@@ -3,6 +3,13 @@ import StrippedWebPlatformUtilsService from './StrippedWebPlatformUtilsService'
 import { CryptoService } from './@bitwarden/jslib/services/crypto.service'
 import { WebCryptoFunctionService } from './@bitwarden/jslib/services/webCryptoFunction.service'
 
+const CIPHERS_DOCTYPE = 'com.bitwarden.ciphers'
+const SETTINGS_DOCTYPE = 'io.cozy.settings'
+
+const isForbiddenError = rawError => {
+  return rawError.message.match(/code=403/)
+}
+
 /**
  * Check if the parameter is an instance URL or an email
  * @param {string} instance_or_email - instance url or email
@@ -82,4 +89,61 @@ function getLightCryptoService() {
     cryptoFunctionService
   )
   return cryptoService
+}
+
+/**
+ * Check if there are ciphers in the database
+ * @param {object} cozyClient - cozy-client instance
+ * @return {bool} whether there are ciphers in the databse or not
+ */
+export const checkHasCiphers = async cozyClient => {
+  try {
+    const { data: ciphers } = await cozyClient.query(
+      cozyClient.find(CIPHERS_DOCTYPE)
+    )
+
+    return ciphers.length > 0
+  } catch (err) {
+    /* eslint-disable no-console */
+    if (isForbiddenError(err)) {
+      console.error(
+        `Your app must have the GET permission on the ${CIPHERS_DOCTYPE} doctype.`
+      )
+    } else {
+      console.error(err.message)
+    }
+    /* eslint-enable no-console */
+
+    return false
+  }
+}
+
+/**
+ * Check if the extension is installed, based on the Cozy settings
+ * @param {object} cozyClient - cozy-client instance
+ * @return {bool} whether the extension is installed or not
+ */
+export const checkHasInstalledExtension = async cozyClient => {
+  try {
+    const { rows: docs } = await cozyClient
+      .getStackClient()
+      .fetchJSON('GET', `/data/${SETTINGS_DOCTYPE}/_normal_docs`)
+
+    const [bitwardenSettings] = docs.filter(
+      doc => doc._id === 'io.cozy.settings.bitwarden'
+    )
+    return bitwardenSettings && !!bitwardenSettings.extension_installed
+  } catch (err) {
+    /* eslint-disable no-console */
+    if (isForbiddenError(err)) {
+      console.error(
+        `Your app must have the GET permission on the ${SETTINGS_DOCTYPE} doctype.`
+      )
+    } else {
+      console.error(err.message)
+    }
+    /* eslint-enable no-console */
+
+    return false
+  }
 }
