+++ src/components/VaultUnlocker.jsx
@@ -6,7 +6,7 @@ import localesEn from '../locales/en.json'
 import localesFr from '../locales/fr.json'
 import Spinner from 'cozy-ui/transpiled/react/Spinner'
 import { withClient } from 'cozy-client'
-import { checkHasCiphers, checkHasInstalledExtension } from '../utils'
+import { checkHasCiphers, checkHasInstalledExtension } from '../CozyUtils'
 
 const locales = {
   en: localesEn,
